+++ ado_wrapper/resources/merge_policies.py
@@ -79,7 +79,9 @@ class MergePolicyDefaultReviewer(StateManagedResource):
                     [x for x in all_reviewers if x.member_id == reviewer_id][0].is_required = True
         # =====================================================================================
         # Fix local_ids (convert them to origin_ids)
-        local_ids_to_origin_ids = AdoUser._convert_local_ids_to_origin_ids(ado_client, [x.member_id for x in all_reviewers])
+        local_ids_to_origin_ids = AdoUser._convert_local_ids_to_origin_ids(
+            ado_client, [x.member_id for x in all_reviewers]
+        )  # pylint: disable=protected-access
         for reviewer in [x for x in all_reviewers if x.member_id is not None]:
             reviewer.member_id = local_ids_to_origin_ids[reviewer.member_id]
         # =====================================================================================
@@ -91,7 +93,9 @@ class MergePolicyDefaultReviewer(StateManagedResource):
     ) -> None:
         if reviewer_origin_id in [x.member_id for x in cls.get_default_reviewers(ado_client, repo_id, branch_name)]:
             raise ValueError("Reviewer already exists! To update, please remove the reviewer first.")
-        local_id = AdoUser._convert_origin_ids_to_local_ids(ado_client, [reviewer_origin_id])[reviewer_origin_id]
+        local_id = AdoUser._convert_origin_ids_to_local_ids(ado_client, [reviewer_origin_id])[
+            reviewer_origin_id
+        ]  # pylint: disable=protected-access
         payload = {
             "type": {"id": _get_type_id(ado_client, "Required reviewers")},
             "isBlocking": is_required,