+++ ado_wrapper/resources/projects.py
@@ -117,6 +117,75 @@ class Project(StateManagedResource):
         return {key: value["enabled"] for key, value in mapping.items() if isinstance(value, dict)}
 
     @staticmethod
+    def get_build_queue_settings(ado_client: "AdoClient") -> dict[str, Any]:
+        """A method which returns the amount of free and paid for parallel jobs within AzureDevops.
+        To get the settings for another project, use ado_client.assume_project()."""
+        # Can't use build_hierarchy_payload because it's wrapped in context...
+        PAYLOAD = {"contributionIds": ["ms.vss-build-web.build-queue-hub-data-provider"], "context": {"properties": {
+                   "sourcePage": {"routeId": "ms.vss-admin-web.project-admin-hub-route"}}}}  # fmt: skip
+        # ============================================================================================================
+        private_projects_request = ado_client.session.post(
+            f"https://dev.azure.com/{ado_client.ado_org_name}/_apis/Contribution/dataProviders/query?api-version=7.0-preview.1",
+            json=PAYLOAD,
+        ).json()["data"]["ms.vss-build-web.build-queue-hub-data-provider"]["taskHubLicenseDetails"]
+        # ============================================================================================================
+        PAYLOAD["context"]["properties"]["FetchPublicResourceUsage"] = True  # type: ignore[index]
+        public_projects_request = ado_client.session.post(
+            f"https://dev.azure.com/{ado_client.ado_org_name}/_apis/Contribution/dataProviders/query?api-version=7.0-preview.1",
+            json=PAYLOAD,
+        ).json()["data"]["ms.vss-build-web.build-queue-hub-data-provider"]["resourceUsages"]
+        public_microsoft_hosted = [x["resourceLimit"] for x in public_projects_request if x["resourceLimit"]["isHosted"]][0]
+        public_self_hosted = [x["resourceLimit"] for x in public_projects_request if not x["resourceLimit"]["isHosted"]][0]
+        # ============================================================================================================
+        data = {
+            "private_projects": {
+                "microsoft_hosted": {
+                    "monthly_purchases": private_projects_request["purchasedHostedLicenseCount"],
+                },
+                "self_hosted": {
+                    "free_parallel_jobs": private_projects_request["freeLicenseCount"],
+                    "visual_studio_enterprise_subscribers": private_projects_request["enterpriseUsersCount"],
+                    "monthly_purchases": private_projects_request["purchasedLicenseCount"],
+                },
+            },
+            "public_projects": {
+                "microsoft_hosted": {
+                    "free_parallel_jobs": int(public_microsoft_hosted["resourceLimitsData"]["FreeCount"]),
+                    "monthly_purchases": int(public_microsoft_hosted["resourceLimitsData"]["PurchasedCount"]),
+                },
+                "self_hosted": {
+                    "free_parallel_jobs": (
+                        public_self_hosted["resourceLimitsData"]["FreeCount"]
+                        if int(public_self_hosted["resourceLimitsData"]["FreeCount"]) < 10000
+                        else "Unlimited"
+                    )
+                },
+            },
+        }
+        return data
+
+    @staticmethod
+    def get_retention_policy_settings(ado_client: "AdoClient") -> dict[str, Any]:
+        """Gets a project's Retention policy settings, including how many days to keep releases, to delete the builds, etc."""
+        request = ado_client.session.get(
+            f"https://vsrm.dev.azure.com/{ado_client.ado_org_name}/{ado_client.ado_project_name}/_apis/Release/releasesettings?api-version=7.0-preview.1",
+        ).json()["retentionSettings"]
+        return {
+            "maximum_retention_policy": {
+                "days_to_retain_a_release": request["maximumEnvironmentRetentionPolicy"]["daysToKeep"],
+                "minimum_releases_to_keep": request["maximumEnvironmentRetentionPolicy"]["releasesToKeep"],
+            },
+            "default_retention_policy": {
+                "days_to_retain_a_release": request["defaultEnvironmentRetentionPolicy"]["daysToKeep"],
+                "minimum_releases_to_keep": request["defaultEnvironmentRetentionPolicy"]["releasesToKeep"],
+                "retain_build": request["defaultEnvironmentRetentionPolicy"]["retainBuild"],
+            },
+            "permanently_destory_releases": {
+                "days_to_keep_releases_after_deletion": request["daysToKeepDeletedReleases"],
+            },
+        }
+
+    @staticmethod
     def get_repository_settings(
         ado_client: "AdoClient", project_name: str | None = None
     ) -> dict[ProjectRepositorySettingType, "ProjectRepositorySettings"]:  # fmt: skip